+++ src/main/scala/com/github/vitalsoftware/scalaredox/models/Medication.scala
@@ -121,7 +121,7 @@ object OrderedMedication extends RobustPrimitives
   ID: String,
   Notes: Seq[String] = Seq.empty,
   Medication: OrderedMedication,
-  Indications: Option[BasicCode],
+  Indications: Seq[BasicCode] = Seq.empty,
   Provider: Option[OrderProvider] = None,
   EnteredBy: Option[Provider] = None,
   VerifiedBy: Option[Provider] = None,
@@ -139,8 +139,8 @@ object MedicationOrder extends RobustPrimitives
 @jsonDefaults case class MedicationsMessage(
   Meta: Meta,
   Patient: Patient,
-  Visit: Visit,
+  Visit: Option[VisitInfo] = None,
   Order: MedicationOrder,
-)
+) extends MetaLike with HasPatient with HasVisitInfo
 
 object MedicationsMessage extends RobustPrimitives